requirements.txt
pandas
sqlalchemy
yfinance
python-dotenv
psycopg2-binary
